requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
cmake==3.21.0
colorama==0.4.4
Django==3.2.5
django-cors-headers==3.7.0
django-email-users==1.1.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-simple-captcha==0.5.14
dnspython==1.16.0
eventlet==0.31.1
Flask==2.0.1
future==0.18.2
greenlet==1.1.0
hog==0.1.7
idna==3.2
IMAPClient==2.2.0
imutils==0.5.4
itchat==1.2.32
itsdangerous==2.0.1
Jinja2==3.0.1
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
MyApplication==0.1.0
numpy==1.21.1
Pillow==8.3.1
pygame==2.0.1
PyMySQL==1.0.2
pypng==0.0.20
PyQRCode==1.2.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
Student==0.0.1
urllib3==1.26.6
Werkzeug==2.0.1
wxpy==0.3.9.8
xlwt==1.3.0
